change path and rename propTypes
